+++ app/src/main/java/ru/surf/learn2invest/ui/components/screens/fragments/profile/ProfileFragmentViewModel.kt
@@ -3,7 +3,7 @@ package ru.surf.learn2invest.ui.components.screens.fragments.profile
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import ru.surf.learn2invest.data.database_components.DatabaseRepository
-import ru.surf.learn2invest.noui.cryptography.FingerprintAuthenticator
+import ru.surf.learn2invest.data.cryptography.FingerprintAuthenticator
 import javax.inject.Inject
 
 @HiltViewModel
