+++ app/src/main/java/ru/surf/learn2invest/ui/components/screens/trading_password/TradingPasswordActivityViewModel.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import ru.surf.learn2invest.R
 import ru.surf.learn2invest.data.database_components.DatabaseRepository
-import ru.surf.learn2invest.noui.cryptography.PasswordHasher
+import ru.surf.learn2invest.data.cryptography.PasswordHasher
 import javax.inject.Inject
 
 @HiltViewModel
