+++ data/build.gradle.kts
@@ -60,4 +60,7 @@ dependencies {
     //coil // TODO (mek implementation)
     api(libs.coil.svg)
     api(libs.coil)
+    // using fingerprint
+    implementation(libs.androidx.biometric)
+    implementation(libs.biometric)
 }
